Make requested changes, clean up logging
