+++ services/app-web/src/hooks/useTealium.ts
@@ -7,7 +7,9 @@ import type {
     TealiumAlertImpressionObject,
     TealiumInlineErrorObject,
     TealiumRadioButtonEventObject,
-    TealiumCheckboxEventObject
+    TealiumCheckboxEventObject,
+    TealiumAccordionEventObject,
+    TealiumFormSubmitEventObject
 } from '../tealium'
 import { TealiumContext } from '../contexts/TealiumContext';
 import {getRouteName} from '../routeHelpers';
@@ -19,6 +21,9 @@ type UseTealiumHookType = {
         tealiumData: Omit<TealiumButtonEventObject, 'event_name'>,
     ) => void,
     logInternalLinkEvent: (
+        tealiumData: TealiumLinkEventObject,
+    ) => void,
+    logExternalLinkEvent: (
         tealiumData: TealiumLinkEventObject,
     ) => void,
     logDropdownSelectionEvent: (
@@ -39,6 +44,12 @@ type UseTealiumHookType = {
     logCheckboxEvent: (
         tealiumData: TealiumCheckboxEventObject
     ) => void
+    logAccordionEvent: (
+        tealiumData: TealiumAccordionEventObject
+    ) => void
+    logFormSubmitEvent: (
+        tealiumData: TealiumFormSubmitEventObject
+    ) => void
 }
 
 const useTealium = (): UseTealiumHookType => {
@@ -49,12 +60,16 @@ const useTealium = (): UseTealiumHookType => {
         return {
             logButtonEvent: warnNoTealium,
             logInternalLinkEvent: warnNoTealium,
+            logExternalLinkEvent: warnNoTealium,
             logDropdownSelectionEvent: warnNoTealium,
             logFilterEvent: warnNoTealium,
             logInlineErrorEvent: warnNoTealium,
             logAlertImpressionEvent: warnNoTealium,
             logRadioButtonEvent: warnNoTealium,
-            logCheckboxEvent: warnNoTealium
+            logCheckboxEvent: warnNoTealium,
+            logFormSubmitEvent: warnNoTealium,
+            logAccordionEvent: warnNoTealium
+
         };
     }
 
@@ -65,7 +80,7 @@ const useTealium = (): UseTealiumHookType => {
     ) => {
         const logData: TealiumButtonEventObject = {
             ...tealiumData,
-            link_type: tealiumData.link_url ? 'link_other' : undefined,
+            link_type:  'link_other',
             event_name: 'button_engagement',
         }
         logUserEvent(logData, pathname, loggedInUser, heading)
@@ -76,7 +91,19 @@ const useTealium = (): UseTealiumHookType => {
     ) => {
         const logData: TealiumLinkEventObject = {
             ...tealiumData,
+            link_type:  'link_other',
             event_name: tealiumData.event_name ?? 'internal_link_clicked',
+        }
+        logUserEvent(logData, pathname, loggedInUser, heading)
+    }
+
+    const logExternalLinkEvent = (
+        tealiumData: TealiumLinkEventObject
+    ) => {
+        const logData: TealiumLinkEventObject = {
+            ...tealiumData,
+            link_type:  'link_external',
+            event_name: 'external_link_click',
         }
         logUserEvent(logData, pathname, loggedInUser, heading)
     }
@@ -101,6 +128,7 @@ const useTealium = (): UseTealiumHookType => {
         const logData: TealiumInlineErrorObject = {
             ...tealiumData,
             event_name: 'inline_error',
+            link_type:  'link_other',
         }
         logUserEvent(logData, pathname, loggedInUser, heading)
     }
@@ -123,6 +151,7 @@ const useTealium = (): UseTealiumHookType => {
     ) => {
         const logData: TealiumRadioButtonEventObject = {
             ...tealiumData,
+            link_type: 'link_other',
             // Alerts usually are placed at top of pages. The way the app works, the form would be closely tied to the
             // pathname, so we can default to using that for the heading.
             event_name: 'radio_button_list_selected',
@@ -134,15 +163,41 @@ const useTealium = (): UseTealiumHookType => {
         tealiumData: TealiumCheckboxEventObject
     ) => logUserEvent(tealiumData, pathname, loggedInUser, heading)
 
+    const logFormSubmitEvent = (
+        tealiumData: TealiumFormSubmitEventObject
+    ) =>  {
+        const logData: TealiumFormSubmitEventObject = {
+            ...tealiumData,
+            link_type: 'link_other',
+            event_name: 'form_field_submit'
+        }
+        logUserEvent(logData, pathname, loggedInUser, heading)
+    }
+
+
+    const logAccordionEvent = (
+        tealiumData: TealiumAccordionEventObject
+    ) => {
+        const logData: TealiumAccordionEventObject = {
+            ...tealiumData,
+            link_type: 'link_other',
+        }
+        logUserEvent(logData, pathname, loggedInUser, heading)
+    }
+
+
     return {
         logButtonEvent,
         logInternalLinkEvent,
+        logExternalLinkEvent,
         logDropdownSelectionEvent,
         logFilterEvent,
         logInlineErrorEvent,
         logAlertImpressionEvent,
         logRadioButtonEvent,
-        logCheckboxEvent
+        logCheckboxEvent,
+        logFormSubmitEvent,
+        logAccordionEvent
     }
 }
