+++ services/app-web/src/pages/CMSDashboard/RateReviewsDashboard/RateReviewsTable.tsx
@@ -34,6 +34,7 @@ import { FilterDateRangeRef } from '../../../components/FilterAccordion/FilterDa
 import { Loading, NavLinkWithLogging } from '../../../components'
 import { useTealium } from '../../../hooks'
 import useDeepCompareEffect from 'use-deep-compare-effect'
+import { getTealiumFiltersChanged } from '../../../tealium/tealiumHelpers'
 
 type RatingPeriodFilterType = [string, string] | []
 
@@ -460,7 +461,9 @@ export const RateReviewsTable = ({
     }, [defaultFiltersFromUrl, defaultColumnFilters])
 
     useDeepCompareEffect(() => {
-        const filterCategories = columnFilters.map((f) => f.id).join(',')
+        const filterCategories = columnFilters.map((f) => f.id)
+        .join(',')
+        const filterCategoriesForAnalytics = getTealiumFiltersChanged(columnFilters)
         const prevFilterCategories = prevFilters.filters
             .map((f) => f.id)
             .join(',')
@@ -474,7 +477,7 @@ export const RateReviewsTable = ({
                 logFilterEvent({
                     event_name: 'filter_removed',
                     search_result_count: submissionCount,
-                    filter_categories_used: filterCategories,
+                    filter_categories_used:  filterCategoriesForAnalytics,
                 })
                 // If there are filters, then we applied new filters
             } else if (columnFilters.length > 0) {
@@ -484,7 +487,7 @@ export const RateReviewsTable = ({
                     results_count_after_filtering: submissionCount,
                     results_count_prior_to_filtering:
                         prevFilters.results ?? 'No prior count, filter on load',
-                    filter_categories_used: filterCategories,
+                    filter_categories_used: filterCategoriesForAnalytics,
                 })
             }
             setPrevFilters({
